+++ packages/cli/src/commands/generate/service/rest.ts
@@ -4,7 +4,7 @@ import { fileURLToPath } from "url";
 import pkg from "fs-extra";
 import { Client } from "../../../util/client.js";
 
-const { copy, existsSync, ensureDir, readJson, writeJson, readFile, writeFile } = pkg;
+const { copy, existsSync, ensureDir, readJson, writeJson, readFile, writeFile, move } = pkg;
 
 export const generateRestService = async (client: Client, name: string, description: string) => {
   const { output } = client;
@@ -14,8 +14,8 @@ export const generateRestService = async (client: Client, name: string, descript
   const serviceName = name.toLowerCase().replace(/\s+/g, "-");
 
   try {
-    const backendTemplate = path.resolve(__dirname, "../../../../src/templates/backend", "rest");
-    const frontendTemplate = path.resolve(__dirname, "../../../../src/templates/frontend", "rest");
+    const backendTemplate = path.resolve(__dirname, "../../../../dist/templates/backend", "rest");
+    const frontendTemplate = path.resolve(__dirname, "../../../../dist/templates/frontend", "rest");
     if (
       existsSync("./services") &&
       existsSync("./apps") &&
@@ -29,16 +29,20 @@ export const generateRestService = async (client: Client, name: string, descript
         .replace("<%= appName %>", serviceName)
         .replace("<%= description %>", description);
       await writeFile(`./services/${serviceName}/package.json`, modifiedTemplateContents);
+      const cdkContents = await readFile(`./services/${serviceName}/cdk/src/index.ts`, "utf-8");
+      const modifiedCdkContents = cdkContents.replace("Service Name", serviceName);
+      await writeFile(`./services/${serviceName}/cdk/src/index.ts`, modifiedCdkContents);
 
-      //frontend
-      await copy(frontendTemplate, "./apps");
-      await writeToPackageJson("./apps/client/package.json", frontendPackageJson);
-      const codegenConfigContents = await readFile("./apps/codegen.ts", "utf-8");
+      // frontend
+      await copy(frontendTemplate, "./apps/client");
+      await move("./apps/client/index.tsx", "./apps/client/pages/index.tsx", { overwrite: true });
+      await writeToPackageJson("./apps/client/package.json");
+      const codegenConfigContents = await readFile("./apps/client/codegen.ts", "utf-8");
       const modifiedCodegenConfigContents = codegenConfigContents.replace(
         "Service Name",
         serviceName,
       );
-      await writeFile("./apps/codegen.ts", modifiedCodegenConfigContents);
+      await writeFile("./apps/client/codegen.ts", modifiedCodegenConfigContents);
       return { success: true };
     } else {
       throw new Error("services and apps folders do not exist");
@@ -49,7 +53,7 @@ export const generateRestService = async (client: Client, name: string, descript
   }
 };
 
-const writeToPackageJson = async (filePath: string, data: any) => {
+const writeToPackageJson = async (filePath: string) => {
   try {
     const packageJson = await readJson(filePath);
 
@@ -75,6 +79,7 @@ const frontendPackageJson = {
   },
   dependencies: {
     "@apollo/client": "^3.7.14",
+    "@graphql-codegen/typescript-operations": "^4.0.0",
     "@graphql-codegen/typescript-react-apollo": "^3.3.7",
     "@graphql-codegen/typescript-resolvers": "^3.2.1",
     "aws-appsync-auth-link": "^3.0.7",