+++ packages/cli/src/templates/frontend/rest/codegen.ts
@@ -5,13 +5,18 @@ const serviceName = "Service Name";
 const config: CodegenConfig = {
   overwrite: true,
   schema: [
-    `../services/${serviceName}/prisma/generated/prisma-appsync/schema.gql`,
+    `../../services/${serviceName}/prisma/generated/prisma-appsync/schema.gql`,
     "./graphql/aws.gql",
   ],
   generates: {
     "./graphql/graphql-types.ts": {
-      plugins: ["typescript", "typescript-resolvers", "typescript-react-apollo"],
-      documents: "src/**/*.gql",
+      plugins: [
+        "typescript",
+        "typescript-operations",
+        "typescript-resolvers",
+        "typescript-react-apollo",
+      ],
+      documents: "graphql/*.gql",
       config: {
         scalars: {
           AWSDateTime: "Date",